refactor(api): give POST /runner/runs/:runId/complete a domain module
The largest handler in the file: a pre-transaction principal read, then one
ReadCommitted transaction of ~780 lines and 33 writes, then a lease release
and an ownership assertion. Only the last two are about the process rather
than the run, and only those stay at the route.

`completeRun(db, { runId, body, claimantClass })` owns the transaction and
its isolation level — "ReadCommitted lets successor CAS losers observe
count=0 instead of surfacing a serialization failure to runners" is a
statement about run completion, not about HTTP, so it moves with the code it
explains. It returns `RunCompletion`, which is the object the transaction
already built, or a named `CompleteRunRefusal`. The post-transaction
`db.run.findFirst` that told WAITING_INBOX from a stale fence moves behind
that refusal: a caller re-deriving it was re-deriving a distinction the
action had already made.

`app.ts` and `run-completion.ts` both call the merge-tail actions — the
base-drift recovery aggregate, the repair and follow-up cards, the stop
notices, and `handleRegressionCompletion`, which other modules import — so
they move to `merge-tail-actions.ts`. Importing them back out of `app.ts`
would have made a cycle; there is no smaller move that avoids one. This is
mechanics the extraction forces, not a second subject.

`activateMergeTailTarget`, `failureEnvelopeV1Input` and the repair
documentation ordinals had no other caller and move into the module.

Two things found and deliberately left alone:

- `options.ownership.assertHeld()` appears exactly twice in the file and the
  two disagree: `/claim` asserts before its transaction, `/complete` after.
  Changing that changes when a control-plane refusal fires on a live route.
- The `{ error: string }` arm of the claim result has no producer: the claim
  transaction only ever returns a claim or `null`. The route's check for it
  is kept as it was.
