+++ packages/api/src/app.ts
@@ -3,8 +3,6 @@ import { createHash } from "node:crypto";
 import {
   ACTIVE_RUN_STATUSES,
   AssigneeType,
-  activateChainSuccessor,
-  advanceTemplateTask,
   agentArchiveBlocker,
   applyInboxDecision,
   catalogRunnerForModel,
@@ -12,9 +10,7 @@ import {
   CodexServiceTier,
   COMPOUND_IMPLEMENTATION_ASSIGNEE_ERROR_CODE,
   deriveRunConfig,
-  FAILURE_ENVELOPE_VERSION,
   FailureClass,
-  failurePhases,
   GoalStatus,
   enqueueTaskRun,
   InboxStatus,
@@ -38,7 +34,6 @@ import {
   RunnerKind,
   RunnerPreference,
   recomputeSessionUsage,
-  resolveRequeueBase,
   resolveRunBranches,
   runnerFor,
   sessionUsageCost,
@@ -50,7 +45,6 @@ import {
   TaskSource,
   TaskStatus,
   TriggerFireSource,
-  gateQuestion,
   isMergeExecutorRunnerId,
   mechanicalPrincipalRefusal,
   executionModeFor,
@@ -60,14 +54,10 @@ import {
   projectMergeOutcome,
   runOwnsMergeOutcome,
   INTEGRATOR_AGENT_NAME,
-  INTEGRATOR_TEMPLATE_NAME,
   MERGE_INTEGRATOR_KIND,
-  isIntegratorStep,
   latestTargetCorrection,
   loadIntegratorTask,
   observedChainPullRequests,
-  parseMergeResult,
-  recordIntegratorStop,
   requestConfirmationCard,
   resolveChainTarget,
   selectAuthorization,
@@ -77,29 +67,9 @@ import {
   type CandidateActivity,
   type CardRow,
   type DecisionRow,
-  AUTHORITY_RESIGN_DEDUPE_PREFIX,
-  AUTHORITY_RESIGN_OPEN_PREFIX,
-  INDEPENDENT_REVIEW_OPEN_PREFIX,
   isMergeReadinessStep,
-  latestMarker,
-  MAX_AUTHORITY_RESIGN_ROUNDS,
-  MAX_BLOCKING_REVIEW_ROUNDS,
-  MERGE_TAIL_KIND,
-  MergeRecoveryStatus,
   mergeRecoveryPhase,
-  openReviewObligation,
-  parseIndependentReviewDecision,
-  parseRegressionVerdict,
-  parseResolverResult,
-  readMarkerHistory,
-  readMarkers,
-  recoveryContext,
-  writeMarker,
-  RELEASE_AUTHORITY_FILE,
-  LEGACY_PRE_ADJUDICATION_TEMPLATE_PREFIX,
-  type IndependentReviewFinding,
   type MergeRecoveryAttempt,
-  type RecoveryContext,
 } from "@agentos/db";
 import { Hono, type Context } from "hono";
 import { cors } from "hono/cors";
@@ -111,10 +81,8 @@ import { authenticate, principalMayAccess, type Principal } from "./auth.js";
 import { boardCard, chainDisplayByTask, etagFor, etagMatches, serializeUsageCost } from "./board.js";
 import { isValidBranchName } from "./branch-name.js";
 import { chainExecutionOwner } from "./chain-execution-owner.js";
-import { FAILURE_REASON_LIMIT, failureReasonText, truncateFailureReason } from "./failure-reason.js";
+import { FAILURE_REASON_LIMIT, failureReasonText } from "./failure-reason.js";
 import {
-  canonicalOutputRefusal,
-  isCanonicalAgentStep,
   isCanonicalBlindFindingsStep,
   isCanonicalSolFindingsStep,
   persistSessionTaskOutput,
@@ -147,14 +115,9 @@ import {
   stepName,
 } from "./chain.js";
 import {
-  classifyEnvelope,
-  completionSucceeded,
-  externalFailure,
-  failureIsRetryable,
   jsonValue,
   makeDedupeKey,
   normalizeSessionEventValue,
-  retryDelayMs,
   runBudgetCeiling,
 } from "./execution.js";
 import { createArchivedRunNoticeScheduler, noteArchivedQueuedRuns, reconcileDatabaseRuns } from "./reconcile.js";
@@ -177,13 +140,12 @@ import { DirectoryNotEmptyError, InvalidPathError, IsADirectoryError, NotADirect
 import {
   hasActiveRun,
   isLiveStatus,
-  lockTask,
   lockTaskMutationRows,
   reactivationBlocked,
 } from "./task-write.js";
 import { patchTask } from "./task-patch.js";
 import { claimRun } from "./run-claim.js";
-import { openMergeTailStopNotice } from "./merge-tail-actions.js";
+import { completeRun } from "./run-completion.js";
 import { withoutUndefined } from "./without-undefined.js";
 import { versionPayload } from "./version.js";
 
@@ -195,11 +157,6 @@ export interface LiveAppOptions {
   releaseMergeLease?: MergeLeaseReleaser;
 }
 
-type DbTx = Prisma.TransactionClient;
-
-/** Full Assurance regression and the documentation node a repair must reopen. */
-const FULL_REPAIR_DOCUMENTATION_ORDINALS = { regression: 10, documentation: 9 } as const;
-const LEGACY_PRE_ADJUDICATION_REPAIR_DOCUMENTATION_ORDINALS = { regression: 11, documentation: 10 } as const;
 
 const mergeRecoveryProjection = (row: MergeRecoveryAttempt | null) => row ? ({
   id: row.id,
@@ -212,459 +169,6 @@ const mergeRecoveryProjection = (row: MergeRecoveryAttempt | null) => row ? ({
   failureReason: row.failureReason,
   updatedAt: row.updatedAt,
 }) : null;
-
-const baseDriftRecoveryContext = async (
-  tx: DbTx,
-  regressionTaskId: string,
-  recoveryRunId?: string,
-  sourceStopId?: string,
-): Promise<RecoveryContext | null> => {
-  const row = await tx.mergeRecoveryAttempt.findFirst({
-    where: {
-      regressionTaskId,
-      status: { in: [MergeRecoveryStatus.REPAIRING, MergeRecoveryStatus.AWAITING_AUTHORIZATION] },
-      ...(recoveryRunId ? { recoveryRunId } : {}),
-      ...(sourceStopId ? { sourceStopId } : {}),
-    },
-    orderBy: [{ attempt: "desc" }, { id: "desc" }],
-  });
-  return recoveryContext(row);
-};
-
-const stopBaseDriftRecoveryTail = async (
-  tx: DbTx,
-  context: RecoveryContext,
-  phase: "regression" | "independent-review",
-  reason: string,
-): Promise<void> => {
-  const body = `Automatic base-drift recovery ${context.attempt} stopped at ${phase}: ${reason}`;
-  await tx.mergeRecoveryAttempt.update({ where: { id: context.aggregateId }, data: {
-    status: MergeRecoveryStatus.BLOCKED_DOWNSTREAM,
-    failureReason: reason,
-    endedAt: new Date(),
-  } });
-  await tx.task.updateMany({
-    where: { id: { in: [context.regressionTaskId, context.readinessTaskId, context.integratorTaskId] } },
-    data: { status: TaskStatus.REVIEW, failureReason: body },
-  });
-  const dedupeKey = `merge-base-drift-recovery-tail-stop:${context.sourceStopId}:${phase}`;
-  await tx.inboxMessage.upsert({ where: { dedupeKey }, create: {
-    from: "AGENT", taskId: context.regressionTaskId, kind: "TEXT", body, dedupeKey,
-  }, update: {} });
-  const metadata = { ...context, state: "tail-stopped", phase, reason, dedupeKey };
-  for (const taskId of [context.integratorTaskId, context.regressionTaskId]) {
-    await writeMarker(tx, taskId, "baseDriftRecovery", { actorType: "control-plane", body, metadata });
-  }
-};
-
-/**
- * Resolves the agent that repairs what this chain's independent review found:
- * the chain's own fix step assignee, or `senior-dev` when the chain has none.
- */
-const mergeTailFixAgentName = async (
-  tx: DbTx,
-  regression: { projectId: string; chainId: string | null; templateId: string | null } | null,
-): Promise<string> => {
-  if (!regression) return "senior-dev";
-  const fixTask = await tx.task.findFirst({
-    where: {
-      projectId: regression.projectId,
-      chainId: regression.chainId,
-      templateId: regression.templateId,
-      templateStep: { outputKind: "fixed-implementation" },
-    },
-    select: { assigneeAgent: { select: { name: true } } },
-  });
-  return fixTask?.assigneeAgent?.name ?? "senior-dev";
-};
-
-/**
- * Parks one follow-up finding as a backlog card.
- *
- * A follow-up is by contract not a reachable behavioural defect, so it never
- * holds the merge; the card is what keeps it from being lost instead.
- */
-const createReviewFollowUpCard = async (
-  tx: DbTx,
-  input: {
-    projectId: string;
-    repoId: string | null;
-    agentName: string;
-    reviewTaskId: string;
-    headSha: string;
-    finding: IndependentReviewFinding;
-  },
-): Promise<{ taskId: string } | { refusal: string }> => {
-  if (!input.repoId) return { refusal: "independent review task has no repository" };
-  const named = await tx.agent.findFirst({
-    where: { projectId: input.projectId, name: input.agentName },
-    select: { id: true },
-  });
-  if (!named) return { refusal: `follow-up agent ${input.agentName} is absent` };
-  // A card assigned to an agent archived a moment later, or pointing at a
-  // revoked grant, is a card nothing can ever run. Both facts are therefore
-  // taken under the same mutexes the archive and revoke paths take.
-  const agent = await lockAgentRow(tx, named.id);
-  if (!agent || agent.archivedAt) return { refusal: `follow-up agent ${input.agentName} is absent or archived` };
-  const grant = await lockAgentRepoGrant(tx, {
-    projectId: input.projectId, agentId: agent.id, repoId: input.repoId,
-  });
-  if (!grant) return { refusal: `follow-up agent ${input.agentName} has no repository grant` };
-  const task = await tx.task.create({ data: {
-    projectId: input.projectId,
-    repoId: input.repoId,
-    name: `Merge tail follow-up: ${input.finding.title}`,
-    description: [
-      input.finding.detail,
-      `Raised as a follow-up by the autonomous merge tail independent review ${input.reviewTaskId} at exact head ${input.headSha}. It did not block that merge.`,
-    ].join("\n\n"),
-    assigneeType: "AGENT",
-    assigneeAgentId: agent.id,
-    approvalGate: false,
-    opensPullRequest: false,
-    status: TaskStatus.BACKLOG,
-  } });
-  await writeMarker(tx, task.id, "reviewObligation", {
-    actorType: "control-plane",
-    body: `Follow-up finding from independent review ${input.reviewTaskId} at ${input.headSha}`,
-    metadata: {
-      state: "follow-up",
-      reviewTaskId: input.reviewTaskId,
-      headSha: input.headSha,
-      title: input.finding.title,
-    },
-  });
-  return { taskId: task.id };
-};
-
-/**
- * The one operator action the autonomous merge tail asks for, and the reason it
- * cannot ask an agent instead: `release-authority.json` is signed with a key
- * that lives outside every checkout, so a chain that moves an attested
- * release-path file can only be unblocked by whoever holds it.
- *
- * The message is written to be run, not interpreted: the two recorded SHAs come
- * out of the attestation already in the checkout, so nothing here has to be
- * filled in by hand. Nothing else in the chain is blocked while it waits, and
- * the resign worker resumes this step on its own once the signature is pushed.
- */
-/**
- * A branch name may legally contain `$`, a backtick or a semicolon, and this
- * message is written to be pasted into a shell that will be holding the release
- * signing key. Every interpolated value is quoted, including the embedded single
- * quote a refname may also carry.
- */
-const shellQuote = (value: string): string => `'${value.replaceAll("'", "'\\''")}'`;
-
-const openAuthorityResignNotice = async (
-  tx: DbTx,
-  input: {
-    taskId: string;
-    agentId: string;
-    sessionId?: string;
-    branch: string;
-    headSha: string;
-    summary: string;
-    round: number;
-  },
-): Promise<void> => {
-  const body = [
-    `Autonomous merge tail: ${RELEASE_AUTHORITY_FILE} must be re-signed before this chain can merge.`,
-    `Branch ${input.branch} moved attested release-path files at head ${input.headSha}, so the migration preflight refuses this tree and the merge gate cannot pass. The signing key never enters a chain, so this is yours to run (request ${input.round} of ${MAX_AUTHORITY_RESIGN_ROUNDS}).`,
-    `What moved:\n${input.summary}`,
-    [
-      "Run this in the checkout that holds the private revalidation document, with the signing key at hand:",
-      "",
-      `  git fetch origin ${shellQuote(input.branch)}`,
-      `  git switch --detach ${shellQuote(input.headSha)}`,
-      "  RELEASE_AUTHORITY_KEY=~/.agentos-keys/release-authority.ed25519 \\",
-      `  GOAL5A0_MASTER_SHA="$(node -p "require('./${RELEASE_AUTHORITY_FILE}').masterSha")" \\`,
-      `  GOAL5A0_CONTROL_PLANE_A_SHA="$(node -p "require('./${RELEASE_AUTHORITY_FILE}').controlPlaneASha")" \\`,
-      "    npm run snapshot:authority",
-      `  npm run db:authority-check -w @agentos/db`,
-      `  git add ${RELEASE_AUTHORITY_FILE}`,
-      `  git commit -m ${shellQuote(`chore(release): re-sign the release authority for ${input.branch}`)}`,
-      `  git push origin ${shellQuote(`HEAD:${input.branch}`)}`,
-    ].join("\n"),
-    `Nothing else is needed. The server watches the pull request and re-runs regression verification as soon as the re-signed ${RELEASE_AUTHORITY_FILE} is on ${input.branch}.`,
-  ].join("\n\n");
-  await tx.inboxMessage.upsert({
-    where: { dedupeKey: `${AUTHORITY_RESIGN_DEDUPE_PREFIX}${input.taskId}:${input.headSha}` },
-    create: {
-      from: "AGENT",
-      agentId: input.agentId,
-      ...(input.sessionId ? { sessionId: input.sessionId } : {}),
-      taskId: input.taskId,
-      kind: "TEXT",
-      body,
-      dedupeKey: `${AUTHORITY_RESIGN_DEDUPE_PREFIX}${input.taskId}:${input.headSha}`,
-    },
-    update: {},
-  });
-};
-
-const createMergeTailRepairTask = async (
-  tx: DbTx,
-  input: {
-    regressionTask: { id: string; projectId: string; repoId: string | null; templateId: string | null; chainId: string | null; targetBranch: string | null };
-    sourceRun: { id: string; branch: string | null };
-    agentName: string;
-    repairKind: "refresh-conflict" | "gate-fix" | "review-fix";
-    headSha: string;
-    baseHeadSha: string;
-    summary: string;
-    now: Date;
-  },
-): Promise<{ taskId: string } | { refusal: string }> => {
-  const { regressionTask } = input;
-  if (!regressionTask.repoId || !regressionTask.chainId || !regressionTask.templateId || !input.sourceRun.branch) {
-    return { refusal: "repair task cannot resolve its chain repository and shared branch" };
-  }
-  const agent = await tx.agent.findFirst({
-    where: { projectId: regressionTask.projectId, name: input.agentName, archivedAt: null },
-  });
-  if (!agent) return { refusal: `required repair agent ${input.agentName} is absent or archived` };
-  const grant = await tx.agentRepoAccess.findFirst({
-    where: { projectId: regressionTask.projectId, agentId: agent.id, repoId: regressionTask.repoId },
-  });
-  if (!grant) return { refusal: `required repair agent ${input.agentName} has no repository grant` };
-
-  const prompt = input.repairKind === "refresh-conflict"
-    ? [
-      `Resolve the refresh conflict between chain head ${input.headSha} and target head ${input.baseHeadSha}.`,
-      input.summary,
-      `Re-run the merge, preserve both intents under the merge-resolver role contract, commit the resolution, and persist the role's versioned JSON bound to start ${input.headSha} and target ${input.baseHeadSha}.`,
-    ].join("\n\n")
-    : [
-      `Repair the autonomous merge tail failure at ${input.headSha} against target ${input.baseHeadSha}.`,
-      input.summary,
-      "Make exactly the changes needed to close this failure, run affected suites, commit, and persist the result as task output.",
-    ].join("\n\n");
-  const task = await tx.task.create({ data: {
-    projectId: regressionTask.projectId,
-    repoId: regressionTask.repoId,
-    name: `Autonomous merge tail: ${input.repairKind}`,
-    description: prompt,
-    assigneeType: "AGENT",
-    assigneeAgentId: agent.id,
-    approvalGate: false,
-    opensPullRequest: false,
-    status: TaskStatus.TODO,
-    targetBranch: input.sourceRun.branch,
-    maxSessionsPerTask: 1,
-  } });
-  const repairRun = await enqueueTaskRun(tx, task.id, input.now);
-  await tx.run.update({
-    where: { id: repairRun.id },
-    data: { branch: input.sourceRun.branch, targetBranch: input.sourceRun.branch },
-  });
-  await writeMarker(tx, regressionTask.id, "repairAttempt", {
-    actorType: "control-plane",
-    body: `Automatic ${input.repairKind} attempt queued at chain head ${input.headSha} against ${input.baseHeadSha}`,
-    metadata: {
-      repairKind: input.repairKind,
-      repairTaskId: task.id,
-      headSha: input.headSha,
-      baseHeadSha: input.baseHeadSha,
-    },
-  });
-  await writeMarker(tx, task.id, "repairAttempt", {
-    actorType: "control-plane",
-    body: `Automatic ${input.repairKind} attempt for regression task ${regressionTask.id}`,
-    metadata: {
-      repairKind: input.repairKind,
-      regressionTaskId: regressionTask.id,
-      headSha: input.headSha,
-      baseHeadSha: input.baseHeadSha,
-    },
-  });
-  await tx.task.update({
-    where: { id: regressionTask.id },
-    data: { status: TaskStatus.REVIEW, failureReason: `${input.repairKind}: automatic repair ${task.id} queued at ${input.headSha}` },
-  });
-  return { taskId: task.id };
-};
-
-export const handleRegressionCompletion = async (
-  tx: DbTx,
-  input: {
-    task: { id: string; projectId: string; repoId: string | null; templateId: string | null; chainId: string | null; targetBranch: string | null };
-    run: { id: string; agentId: string; branch: string | null; headSha: string | null; sessionId: string };
-    now: Date;
-  },
-): Promise<"advance" | "handled"> => {
-  const persistedOutput = await tx.taskStepOutput.findUnique({ where: { taskId: input.task.id } });
-  // Regression evidence is run-scoped even though TaskStepOutput is not yet.
-  // An earlier attempt's explicit verdict is not this attempt's output and must
-  // not be reused when the current agent finishes without calling task_output.
-  const output = persistedOutput?.runId === input.run.id ? persistedOutput : null;
-  const recovery = await baseDriftRecoveryContext(tx, input.task.id, input.run.id);
-  const parsed = parseRegressionVerdict(output?.body);
-  const stop = async (reason: string): Promise<"handled"> => {
-    if (recovery) {
-      await stopBaseDriftRecoveryTail(tx, recovery, "regression", reason);
-      return "handled";
-    }
-    await tx.task.update({ where: { id: input.task.id }, data: { status: TaskStatus.REVIEW, failureReason: reason } });
-    await writeMarker(tx, input.task.id, "regression", {
-      actorType: "control-plane",
-      body: `Regression did not advance: ${reason}`,
-      metadata: { state: "stopped", reason },
-    });
-    await openMergeTailStopNotice(tx, { taskId: input.task.id, agentId: input.run.agentId, sessionId: input.run.sessionId, reason });
-    return "handled";
-  };
-  if (parsed.status === "invalid") return stop(parsed.reason);
-  const verdict = parsed.verdict;
-  const effectiveHead = input.run.headSha ?? output?.commitSha ?? null;
-  if (effectiveHead !== verdict.headSha || output?.commitSha !== verdict.headSha) {
-    return stop(`stale regression evidence: verdict ${verdict.headSha}, output ${output?.commitSha ?? "missing"}, run ${effectiveHead ?? "missing"}`);
-  }
-  await writeMarker(tx, input.task.id, "regression", {
-    actorType: "control-plane",
-    body: `Regression ${verdict.outcome} recorded for chain head ${verdict.headSha} against target ${verdict.baseHeadSha}`,
-    metadata: { ...verdict },
-  });
-  if (verdict.outcome === "pass") {
-    if (recovery) {
-      await tx.mergeRecoveryAttempt.update({ where: { id: recovery.aggregateId }, data: {
-        status: MergeRecoveryStatus.AWAITING_AUTHORIZATION,
-        failureReason: null,
-      } });
-    }
-    return "advance";
-  }
-
-  if (verdict.outcome === "authority-resign") {
-    // Rounds are counted from the notices this function itself wrote, not from
-    // the activity log: `actorType` and `metadata` are caller-supplied on the
-    // activity route, and an inbox `dedupeKey` is written nowhere else. The
-    // bound is anti-thrash — a repeat means the last signature did not cover
-    // this tree, and a chain that keeps asking is not progressing.
-    const round = await tx.inboxMessage.count({
-      where: { taskId: input.task.id, dedupeKey: { startsWith: `${AUTHORITY_RESIGN_DEDUPE_PREFIX}${input.task.id}:` } },
-    }) + 1;
-    const exhausted = round > MAX_AUTHORITY_RESIGN_ROUNDS;
-    const branch = input.run.branch;
-    if (exhausted || !branch) {
-      const reason = exhausted
-        ? `release authority re-signature round ${round} exceeds ${MAX_AUTHORITY_RESIGN_ROUNDS} at ${verdict.headSha}: ${verdict.summary}`
-        : `release authority re-signature is required at ${verdict.headSha} but this run names no chain branch`;
-      const bounded = truncateFailureReason(reason, FAILURE_REASON_LIMIT);
-      if (recovery) {
-        await stopBaseDriftRecoveryTail(tx, recovery, "regression", bounded);
-        return "handled";
-      }
-      return stop(bounded);
-    }
-    // A step an operator has already taken over is not re-parked under it: the
-    // completion holds the chain mutex only from here, so the row is claimed
-    // rather than overwritten.
-    const parked = await tx.task.updateMany({
-      where: { id: input.task.id, status: { in: [TaskStatus.TODO, TaskStatus.DOING] } },
-      data: { status: TaskStatus.REVIEW, failureReason: `${AUTHORITY_RESIGN_OPEN_PREFIX}${verdict.headSha}` },
-    });
-    if (parked.count !== 1) {
-      await writeMarker(tx, input.task.id, "authorityResign", {
-        actorType: "control-plane",
-        body: `Release authority re-signature is required at ${verdict.headSha} but this step is no longer the tail's to park`,
-        metadata: {
-          state: "park-skipped",
-          headSha: verdict.headSha,
-          summary: verdict.summary,
-        },
-      });
-      return "handled";
-    }
-    await writeMarker(tx, input.task.id, "authorityResign", {
-      actorType: "control-plane",
-      body: `Release authority re-signature requested at ${verdict.headSha}: ${verdict.summary}`,
-      metadata: {
-        state: "open",
-        headSha: verdict.headSha,
-        baseHeadSha: verdict.baseHeadSha,
-        branch,
-        round,
-        summary: verdict.summary,
-      },
-    });
-    await openAuthorityResignNotice(tx, {
-      taskId: input.task.id,
-      agentId: input.run.agentId,
-      sessionId: input.run.sessionId,
-      branch,
-      headSha: verdict.headSha,
-      summary: verdict.summary,
-      round,
-    });
-    // A base-drift recovery that reaches this needs the same signature as any
-    // other chain, and the resign worker resumes the same step for it. The
-    // recovery aggregate keeps its own state; nothing about it is decided here.
-    return "handled";
-  }
-
-  if (recovery) {
-    await stopBaseDriftRecoveryTail(
-      tx,
-      recovery,
-      "regression",
-      truncateFailureReason(
-        verdict.outcome === "refresh-conflict"
-          ? `refresh conflict at ${verdict.headSha} against ${verdict.baseHeadSha}: ${verdict.summary}`
-          : verdict.outcome === "review-fail"
-            ? `semantic regression FAIL at ${verdict.headSha} against ${verdict.baseHeadSha}: ${verdict.summary}`
-            : `merge gate FAIL at ${verdict.headSha} against ${verdict.baseHeadSha}: ${verdict.summary}`,
-        FAILURE_REASON_LIMIT,
-      ),
-    );
-    return "handled";
-  }
-
-  // The whole history, not the recent-state window: one automatic attempt per
-  // repair kind is the rule, and an attempt pushed past the window by later
-  // activity would license a second one.
-  const attempts = await readMarkerHistory(tx, input.task.id);
-  const repairKind = verdict.outcome === "refresh-conflict"
-    ? "refresh-conflict"
-    : verdict.outcome === "review-fail" ? "review-fix" : "gate-fix";
-  const alreadyAttempted = attempts.some((marker) => (
-    marker.kind === "repairAttempt"
-    && marker.repairKind === repairKind
-    && (repairKind !== "refresh-conflict" || marker.headSha === verdict.headSha)
-  ));
-  if (alreadyAttempted) {
-    return stop(repairKind === "refresh-conflict"
-      ? `second refresh conflict on chain head ${verdict.headSha}`
-      : repairKind === "review-fix"
-        ? `second semantic regression FAIL on chain head ${verdict.headSha}`
-        : `second merge gate FAIL on chain head ${verdict.headSha}`);
-  }
-  let agentName = "merge-resolver";
-  if (repairKind === "gate-fix" || repairKind === "review-fix") {
-    const fixTask = await tx.task.findFirst({
-      where: {
-        projectId: input.task.projectId,
-        chainId: input.task.chainId,
-        templateId: input.task.templateId,
-        templateStep: { outputKind: "fixed-implementation" },
-      },
-      select: { assigneeAgent: { select: { name: true } } },
-    });
-    agentName = fixTask?.assigneeAgent?.name ?? "senior-dev";
-  }
-  const repair = await createMergeTailRepairTask(tx, {
-    regressionTask: input.task,
-    sourceRun: input.run,
-    agentName,
-    repairKind,
-    headSha: verdict.headSha,
-    baseHeadSha: verdict.baseHeadSha,
-    summary: verdict.summary,
-    now: input.now,
-  });
-  if ("refusal" in repair) return stop(repair.refusal);
-  return "handled";
-};
 
 /**
  * An approval gate's `taskId` is the gate step itself; the artifact the approver
@@ -1033,23 +537,6 @@ const versionedEnvelopeInput = z.object({
 // envelope must never be the reason a completion fails — which is also why the
 // route below treats a v1 envelope that fails this schema as no envelope at
 // all rather than as a bad request.
-const failureEnvelopeV1Input = z.object({
-  version: z.number().int().positive(),
-  phase: z.enum(failurePhases),
-  runnerClass: z.nativeEnum(FailureClass).nullable().default(null),
-  exitCode: z.number().int().nullable().default(null),
-  signal: z.string().max(64).nullable().default(null),
-  terminationReason: z.string().max(4000).nullable().default(null),
-  terminalEventSeen: z.boolean().default(false),
-  terminalSuccess: z.boolean().default(false),
-  agentExited: z.boolean().default(false),
-  providerError: z.string().max(64_000).nullable().default(null),
-  stderrSummary: z.string().max(64_000).nullable().default(null),
-  stdoutSummary: z.string().max(64_000).nullable().default(null),
-  timedOut: z.boolean().default(false),
-  transient: z.boolean().default(false),
-  timeoutMs: z.number().int().nonnegative().nullable().default(null),
-});
 
 const completionInput = z.object({
   runnerId: z.string().trim().min(1).max(120),
@@ -1083,6 +570,7 @@ const completionInput = z.object({
   workspaceRetained: z.boolean().default(false),
   failureEnvelope: versionedEnvelopeInput.optional(),
 });
+export type CompletionInput = z.infer<typeof completionInput>;
 const eventInput = z.object({
   seq: z.number().int().nonnegative(),
   at: z.coerce.date().optional(),
@@ -1465,121 +953,6 @@ const settleCancellation = async (
   };
 };
 
-/**
- * Merge-tail repair markers point at an existing canonical task rather than a
- * linked-list successor. Queue that explicit target under the same layer mutex
- * as ordinary chain activation; readiness remains server-owned and is only
- * marked queued for its worker.
- */
-const activateMergeTailTarget = async (
-  tx: Prisma.TransactionClient,
-  taskId: string,
-  now: Date,
-): Promise<void> => {
-  if (!await lockTaskMutationRows(tx, taskId)) return;
-  const target = await tx.task.findUnique({
-    where: { id: taskId },
-    include: {
-      runs: { where: { status: { in: ACTIVE_RUN_STATUSES } }, take: 1 },
-      assigneeAgent: { select: { name: true, archivedAt: true } },
-      templateStep: { include: { taskTemplate: { select: { name: true } } } },
-    },
-  });
-  if (!target || target.status === TaskStatus.DONE || target.runs.length > 0) return;
-  if (target.archivedAt) {
-    await tx.taskActivity.create({ data: {
-      taskId,
-      actorType: "control-plane",
-      body: "Merge-tail target is archived and was not queued",
-    } });
-    return;
-  }
-  if (isMergeReadinessStep(target.templateStep)) {
-    // Readiness runs on the server worker, which only claims TODO/DOING. The
-    // obligation parked this step in REVIEW while the review ran, so resolving
-    // the review has to hand it back; anything else in REVIEW is a real stop
-    // and stays stopped.
-    const resumed = await tx.task.updateMany({
-      where: { id: taskId, status: TaskStatus.REVIEW, failureReason: { startsWith: INDEPENDENT_REVIEW_OPEN_PREFIX } },
-      data: { status: TaskStatus.TODO, failureReason: null },
-    });
-    await tx.taskActivity.create({ data: {
-      taskId,
-      actorType: "control-plane",
-      body: resumed.count === 1
-        ? "Independent review resolved; readiness target returned to the server worker"
-        : "Merge-tail readiness target queued for server worker",
-      metadata: { kind: MERGE_TAIL_KIND.readiness, schemaVersion: 1, state: "queued" },
-    } });
-    return;
-  }
-  if (target.status === TaskStatus.REVIEW && target.failureReason?.startsWith(AUTHORITY_RESIGN_OPEN_PREFIX)) {
-    // A park the resign worker owns. Re-queueing this step now would spend a
-    // gate on a tree the migration preflight still refuses, and would lose the
-    // park the operator's inbox message points at.
-    await tx.taskActivity.create({ data: {
-      taskId,
-      actorType: "control-plane",
-      body: "Merge-tail target is held by an open release authority re-signature",
-      metadata: { kind: MERGE_TAIL_KIND.authorityResign, schemaVersion: 1, state: "held" },
-    } });
-    return;
-  }
-  if (target.assigneeAgent?.archivedAt) {
-    const reason = `Assignee ${target.assigneeAgent.name} is archived; unarchive the agent and retry to queue this merge-tail target`;
-    await tx.task.update({ where: { id: taskId }, data: { status: TaskStatus.REVIEW, failureReason: reason } });
-    await tx.taskActivity.create({ data: {
-      taskId,
-      actorType: "control-plane",
-      body: `Merge-tail target not queued because assignee ${target.assigneeAgent.name} is archived`,
-    } });
-    return;
-  }
-  const claimed = await tx.task.updateMany({
-    where: { id: taskId, status: { in: [TaskStatus.TODO, TaskStatus.DOING, TaskStatus.REVIEW] } },
-    data: { status: TaskStatus.TODO, failureReason: null },
-  });
-  if (claimed.count !== 1) return;
-  const rawTx = tx as Prisma.TransactionClient & { $executeRawUnsafe?: (query: string) => Promise<number> };
-  const savepoint = "merge_tail_enqueue";
-  const hasSavepoint = typeof rawTx.$executeRawUnsafe === "function";
-  if (hasSavepoint) await rawTx.$executeRawUnsafe!(`SAVEPOINT ${savepoint}`);
-  try {
-    await enqueueTaskRun(tx, taskId, now);
-    if (hasSavepoint) await rawTx.$executeRawUnsafe!(`RELEASE SAVEPOINT ${savepoint}`);
-  } catch (error: unknown) {
-    if (hasSavepoint) {
-      await rawTx.$executeRawUnsafe!(`ROLLBACK TO SAVEPOINT ${savepoint}`);
-      await rawTx.$executeRawUnsafe!(`RELEASE SAVEPOINT ${savepoint}`);
-    }
-    const duplicateRun = error instanceof Prisma.PrismaClientKnownRequestError && error.code === "P2002";
-    if (!isArchivedAssigneeError(error) && !isArchivedTaskError(error)
-      && !isIntegratorStoppedError(error) && !duplicateRun) throw error;
-    if (duplicateRun) {
-      await tx.taskActivity.create({ data: {
-        taskId,
-        actorType: "control-plane",
-        body: "Merge-tail target already has the run created by a concurrent activation",
-      } });
-      return;
-    }
-    await tx.task.update({
-      where: { id: taskId },
-      data: { status: TaskStatus.REVIEW, failureReason: error instanceof Error ? error.message : "Merge-tail target could not be queued" },
-    });
-    await tx.taskActivity.create({ data: {
-      taskId,
-      actorType: "control-plane",
-      body: `Merge-tail target was not queued: ${error instanceof Error ? error.message : "enqueue refused"}`,
-    } });
-    return;
-  }
-  await tx.taskActivity.create({ data: {
-    taskId,
-    actorType: "control-plane",
-    body: "Merge-tail target queued",
-  } });
-};
 
 /** Locks a whole candidate set in one statement. `ORDER BY "id"` is not
  *  decoration: it is what stops two concurrent Archive All presses from
@@ -4792,804 +4165,16 @@ export const createApp = (db: PrismaClient, options: LiveAppOptions): Hono<AppEn
     const runId = id.parse(context.req.param("runId"));
     const body = await readJson(context.req.raw, completionInput);
     const principal = context.get("principal");
-    const now = new Date();
-    const fence: RunFence = { runId, runnerId: body.runnerId, fencingToken: body.fencingToken, at: now };
-    // §4.0. Completing a mechanical run is what makes the chain believe a merge
-    // happened, so it is bound to the same independently authenticated
-    // principal that was allowed to claim it — and, symmetrically, the executor
-    // credential completes nothing else. Read before the transaction: the
-    // step binding of a claimed run is immutable (§D-P4 refuses to move it), so
-    // there is no state to lose by refusing here, and nothing has been written.
-    const completing = await db.run.findUnique({
-      where: { id: runId },
-      select: { runnerId: true, task: { select: { templateStep: { select: { stepIndex: true, outputKind: true, taskTemplate: { select: { name: true } } } } } } },
+    const result = await completeRun(db, {
+      runId,
+      body,
+      claimantClass: principal.kind === "merge-executor" ? "merge-executor" : "runner",
     });
-    if (completing) {
-      const refusal = mechanicalPrincipalRefusal(
-        executionModeFor(completing.task?.templateStep ?? null),
-        principal.kind === "merge-executor" ? "merge-executor" : "runner",
-        completing.runnerId ?? body.runnerId,
-      );
-      if (refusal) return context.json({ error: refusal }, 403);
-    }
-    const result = await db.$transaction(async (tx) => {
-      // Run owns fencing, cancellation, and terminalization. Take that mutex
-      // before Task so completion, cancellation, and canonical output writes
-      // cannot deadlock by entering the same two rows in opposite orders.
-      await lockRunRow(tx, runId);
-      const run = await tx.run.findFirst({
-        where: fencedRunWhere(fence),
-        include: {
-          // §D-P5. The step's template name is part of the step-12 identity, so
-          // the completion route has to read it rather than the step alone.
-          task: { include: { templateStep: { include: { taskTemplate: { select: { name: true } } } }, repo: { select: { defaultBranch: true } } } },
-          session: true,
-        },
-      });
-      if (!run?.session) return null;
-      const succeeded = completionSucceeded({
-        exitCode: body.exitCode,
-        signal: body.signal ?? null,
-        terminalEventSeen: body.terminalEventSeen,
-        terminalSuccess: body.terminalSuccess,
-        terminationReason: body.terminationReason ?? null,
-      });
-      // The runner is on the untrusted side of this boundary. When it reports a
-      // structured envelope, the API classifies from the facts in it and
-      // ignores the runner's own `failureClass`/`retryable`/`externalFailure`
-      // — before this, `body.retryable ?? …` meant the runner always won and
-      // the retry whitelist in execution.ts was dead code, so a stdout-derived
-      // misverdict of RATE_LIMITED spent the task's whole run budget on retries
-      // that could not succeed.
-      //
-      // A runner too old to send an envelope, or one sending a version this API
-      // does not know, keeps the previous behaviour verbatim. That is the point
-      // of the version check: an unrecognised shape must not be half-read.
-      const known = body.failureEnvelope?.version === FAILURE_ENVELOPE_VERSION
-        ? failureEnvelopeV1Input.safeParse(body.failureEnvelope)
-        : null;
-      const envelope = !succeeded && known?.success ? known.data : null;
-      const verdict = envelope ? classifyEnvelope(envelope) : null;
-      const failureClass = succeeded
-        ? null
-        : verdict?.failureClass ?? body.failureClass ?? (body.exitCode === 0 ? FailureClass.PROTOCOL_ERROR : FailureClass.TASK_FAILED);
-      const retryable = failureClass
-        ? verdict?.retryable ?? body.retryable ?? failureIsRetryable(failureClass)
-        : false;
-      const retryAt = failureClass && retryable ? new Date(now.getTime() + retryDelayMs(run.runNumber, failureClass)) : null;
-      // An external failure buys the task one more attempt rather than spending one.
-      const external = verdict
-        ? verdict.externalFailure
-        : externalFailure({ succeeded, signal: body.signal ?? null, reported: body.externalFailure, failureClass });
-      // §D-P5 / MF-5. For the integrator step that compensation is switched off
-      // entirely, so the answer transaction is the *only* writer of a ceiling
-      // above the task's original. Otherwise a run authorized once could buy
-      // itself unbounded further attempts by failing externally, and "only a
-      // human re-authorization may exceed the ceiling" would be false in a
-      // reachable interleaving rather than merely hard to reach. The failure
-      // envelope's verdict decides *whether* the failure was external; it does
-      // not get to raise the ceiling on this step either.
-      const mechanical = isIntegratorStep(run.task?.templateStep ?? null);
-      const refunded = external && !mechanical ? 1 : 0;
-      const budgetCeiling = run.maxRunsPerTask + refunded;
-      // One marker read for both completion outcomes; this handler used to
-      // declare `tailRows` twice and scan twice. The success path consults it
-      // only for a standalone auxiliary task — an automatic repair or an
-      // independent review, neither of which is a chain step — while the
-      // failure path consults it only for a failure that is not about to be
-      // retried, which is why the ceiling is computed above the read.
-      const failureIsFinal = !succeeded && !(retryable && run.runNumber < budgetCeiling);
-      const tailMarkers = run.task && (failureIsFinal || (succeeded && !run.task.templateId && !run.task.chainId))
-        ? await readMarkers(tx, run.task.id)
-        : [];
-      const succeededMarkers = succeeded ? tailMarkers : [];
-      const repairMarker = latestMarker(succeededMarkers, "repairAttempt");
-      const reviewMarker = openReviewObligation(succeededMarkers);
-      const repairRegression = repairMarker?.regressionTaskId
-        ? await tx.task.findUnique({
-            where: { id: repairMarker.regressionTaskId },
-            select: {
-              projectId: true,
-              chainId: true,
-              templateId: true,
-              chainIndex: true,
-              templateStep: { select: { stepIndex: true, taskTemplate: { select: { name: true } } } },
-            },
-          })
-        : null;
-      const reviewRegression = reviewMarker?.regressionTaskId
-        ? await tx.task.findUnique({
-            where: { id: reviewMarker.regressionTaskId },
-            select: { projectId: true, repoId: true, templateId: true, chainId: true, targetBranch: true },
-          })
-        : null;
-      // The ordinals are the Full Assurance graph's, and the renamed
-      // adjudication-era rows keep the ones they were created under: a repair
-      // that lands on a chain from either graph still has to put its
-      // documentation node back.
-      const repairDocumentationOrdinals = repairRegression?.templateStep?.taskTemplate.name === INTEGRATOR_TEMPLATE_NAME
-        ? FULL_REPAIR_DOCUMENTATION_ORDINALS
-        : repairRegression?.templateStep?.taskTemplate.name.startsWith(LEGACY_PRE_ADJUDICATION_TEMPLATE_PREFIX)
-          ? LEGACY_PRE_ADJUDICATION_REPAIR_DOCUMENTATION_ORDINALS
-          : null;
-      const repairDocumentationTask = repairRegression?.chainId && repairRegression.templateId
-        && repairDocumentationOrdinals
-        && repairRegression.chainIndex === repairDocumentationOrdinals.regression
-        && repairRegression.templateStep?.stepIndex === repairDocumentationOrdinals.regression
-        ? await tx.task.findFirst({
-            where: {
-              projectId: repairRegression.projectId,
-              chainId: repairRegression.chainId,
-              templateId: repairRegression.templateId,
-              chainIndex: repairDocumentationOrdinals.documentation,
-              archivedAt: null,
-              templateStep: { stepIndex: repairDocumentationOrdinals.documentation, outputKind: "documentation" },
-            },
-            orderBy: { chainIndex: "desc" },
-            select: { id: true },
-          })
-        : null;
-      // An auxiliary task is one whose own marker names the Regression it serves.
-      const mergeTailAuxiliary = Boolean(
-        repairMarker?.regressionTaskId || (reviewMarker?.readinessTaskId && reviewMarker.regressionTaskId),
-      );
-      const auxiliaryTargetTaskId = repairMarker?.regressionTaskId
-        ? repairDocumentationTask?.id ?? repairMarker.regressionTaskId
-        : reviewMarker?.readinessTaskId ?? null;
-      // The same refund, recorded apart from the ceiling it produced. The gates
-      // an operator can reach read this rather than `maxRunsPerTask`, because
-      // only this can still be told apart from the configured budget after that
-      // budget changes. The in-flight ceiling stays derived from the run's own
-      // row: a task's budget being edited mid-run must not retroactively refuse
-      // an attempt already authorized.
-      const budgetGrants = run.budgetGrants + refunded;
-      const tailLeaseChainId = run.task?.chainId ?? repairRegression?.chainId ?? reviewRegression?.chainId ?? null;
-      let releaseMergeLeaseTask: string | null = null;
-      // Completion always mutates its Task, including terminal non-retryable
-      // failures. Run is already locked above; acquire the Task/chain mutex now
-      // for every outcome rather than only the branches that may retry or
-      // advance.
-      if (run.task) {
-        if (run.task.chainId) {
-          await lockChainRows(tx, { projectId: run.task.projectId, chainId: run.task.chainId });
-        } else {
-          await lockTask(tx, run.task.id);
-        }
-      }
-      if (auxiliaryTargetTaskId && auxiliaryTargetTaskId !== run.task?.id) {
-        await lockTaskMutationRows(tx, auxiliaryTargetTaskId);
-      }
-      if (run.task && typeof (tx.task as { findUnique?: unknown }).findUnique === "function") {
-        await tx.task.findUnique({ where: { id: run.task.id }, select: { status: true } });
-      }
-      // Keep the status observed with the fenced Run as the compare-and-set
-      // expectation. The locked re-read above supplies current chain state,
-      // but adopting its newer status as the expectation would let completion
-      // overwrite an operator decision that won while completion waited for
-      // the mutex (for example DONE -> REVIEW on a successful standalone run).
-      const completionTaskStatus = run.task?.status;
-      const terminalStatus = succeeded
-        ? RunStatus.SUCCEEDED
-        : body.terminationReason?.includes("walltime") || body.terminationReason?.includes("stall")
-          ? RunStatus.TIMED_OUT
-          : RunStatus.FAILED;
-      const closed = await tx.run.updateMany({
-        // The terminal compare-and-set deliberately drops `runnerId`: settlement
-        // races completion on this row and neither may win twice. Same instant
-        // as the read above, which is what `at` on the fence is for.
-        where: fencedRunWhere({ runId, fencingToken: body.fencingToken, at: now }),
-        data: {
-          status: terminalStatus,
-          endedAt: now,
-          leaseExpiresAt: null,
-          sessionTokenRevokedAt: now,
-          failureClass,
-          failureReason: succeeded ? null : body.failureReason ?? "Execution failed",
-          retryable,
-          retryAt,
-          // Stored whether or not this API understood it: an envelope from a
-          // future runner is still the evidence of what happened, and the
-          // reason a verdict can be re-decided later instead of re-run.
-          failureEnvelope: succeeded || !body.failureEnvelope ? Prisma.DbNull : jsonValue(body.failureEnvelope),
-          // Kept on the run that produced it, whatever became of that run. The
-          // same tail used to reach this route on every completion and be read
-          // only by the step-output synthesis below, which runs for successful
-          // template/chain/follow-up runs and nothing else — so a failure's own
-          // account of itself died in this handler and the incident could only
-          // be guessed at afterwards. A runner too old to send one leaves NULL,
-          // exactly as before.
-          output: body.output ?? null,
-          terminationReason: body.terminationReason ?? null,
-          branch: body.branch ?? run.branch,
-          // Completion is a second publication write, never an eraser of the
-          // immediate post-push ACK recorded on this run.
-          pushedBranch: body.pushedBranch ?? run.pushedBranch,
-          baseSha: body.baseSha ?? run.baseSha,
-          headSha: body.headSha ?? null,
-          pushStatus: body.pushStatus,
-          pushRemote: body.pushRemote ?? null,
-          pushError: body.pushError ?? null,
-          pullRequestUrl: body.pullRequestUrl ?? null,
-          pullRequestNumber: body.pullRequestNumber ?? null,
-          deliveryInstructions: body.deliveryInstructions ?? null,
-          workspaceRetained: body.workspaceRetained,
-          maxRunsPerTask: budgetCeiling,
-          budgetGrants,
-        },
-      });
-      if (closed.count !== 1) return null;
-      await tx.session.update({
-        where: { id: run.session.id },
-        data: {
-          executionStatus: succeeded ? SessionExecutionStatus.SUCCEEDED
-            : terminalStatus === RunStatus.TIMED_OUT ? SessionExecutionStatus.TIMED_OUT : SessionExecutionStatus.FAILED,
-          cleanupStatus: body.cleanupStatus,
-          exitCode: body.exitCode,
-          signal: body.signal ?? null,
-          terminationReason: body.terminationReason ?? null,
-          endedAt: now,
-          cleanupEndedAt: now,
-          failureReason: succeeded ? null : body.failureReason ?? "Execution failed",
-          cleanupFailureReason: body.cleanupFailureReason ?? null,
-        },
-      });
-      let retryCreated = false;
-      if (!succeeded && retryable && run.task && run.runNumber < budgetCeiling) {
-        const currentTask = await tx.task.findUniqueOrThrow({
-          where: { id: run.task.id },
-          include: { templateStep: true, repo: { select: { defaultBranch: true } } },
-        });
-        // The fifth run-creating path. Indexed chains already resolve their
-        // branch here; template chains must do the same or a retry is created
-        // with `branch: null` and workspace.ts silently moves it to a per-run
-        // branch. Pass the failed template run as the prior so publication
-        // evidence — including WIP salvage written by this completion — still
-        // decides the retry's base; the resolved logical chain head wins over
-        // that run's workspace branch. Non-template
-        // chains retain their existing no-prior resolution, and non-chain
-        // retries retain the historical `branch: null` behavior.
-        //
-        // All of this runs *after* the updateMany that writes the completing
-        // run's `branch`/`pushedBranch`, so the run's own push — a chain step's
-        // publication, or a failed run's salvage — is evidence in this
-        // transaction. `body.branch ?? run.branch` is that same effective value,
-        // because `run` was read before the update.
-        const resolveChain = currentTask.repo && currentTask.chainId
-          && (currentTask.templateId || currentTask.chainIndex !== null);
-        const branches = resolveChain && currentTask.repo
-          ? await resolveRunBranches(
-            tx,
-            { ...currentTask, repo: currentTask.repo },
-            currentTask.templateId ? { branch: body.branch ?? run.branch } : null,
-          )
-          : {
-            branch: null,
-            targetBranch: currentTask.repo
-              ? await resolveRequeueBase(tx, { ...currentTask, repo: currentTask.repo }, {
-                branch: body.branch ?? run.branch,
-                targetBranch: run.targetBranch,
-              })
-              : run.targetBranch,
-          };
-        await tx.run.create({
-          data: {
-            projectId: run.projectId,
-            taskId: run.taskId,
-            goalId: run.goalId,
-            agentId: run.agentId,
-            repoId: run.repoId,
-            runNumber: run.runNumber + 1,
-            dedupeKey: makeDedupeKey(run.task.id, run.runNumber + 1),
-            runner: run.runner,
-            model: run.model,
-            codexServiceTier: run.codexServiceTier,
-            subagentModel: run.subagentModel,
-            subagentMaxConcurrent: run.subagentMaxConcurrent,
-            targetBranch: branches.targetBranch,
-            branch: branches.branch,
-            opensPullRequest: currentTask.opensPullRequest,
-            promptHash: run.promptHash,
-            maxDurationMin: run.maxDurationMin,
-            stallTimeoutMin: run.stallTimeoutMin,
-            maxRunsPerTask: budgetCeiling,
-            budgetGrants,
-            readyAt: retryAt ?? now,
-          },
-        });
-        retryCreated = true;
-      }
-      if (!succeeded && !retryCreated && (mechanical
-        || run.task?.templateStep?.outputKind === "regression-verification"
-        || mergeTailAuxiliary)) {
-        releaseMergeLeaseTask = tailLeaseChainId;
-      }
-      if (run.taskId) {
-        const budgetExhausted = !succeeded && retryable && !retryCreated;
-        let canonicalOutputFailure: string | null = null;
-        if (!succeeded && !retryCreated && run.task) {
-          // The same markers the success path above read, from the same scan.
-          // A failed auxiliary task closes the obligation it was carrying, so
-          // its review obligation counts in any state, not only `open`.
-          const failedRepair = latestMarker(tailMarkers, "repairAttempt");
-          const failedReview = latestMarker(tailMarkers, "reviewObligation");
-          if (failedRepair?.regressionTaskId) {
-            const reason = `${failedRepair.repairKind} repair ${run.taskId} failed without closing the repair at ${failedRepair.headSha}`;
-            await tx.task.update({ where: { id: failedRepair.regressionTaskId }, data: { status: TaskStatus.REVIEW, failureReason: reason } });
-            await writeMarker(tx, failedRepair.regressionTaskId, "repairResult", {
-              actorType: "control-plane",
-              body: `Automatic ${failedRepair.repairKind} attempt failed: ${failedRepair.headSha} -> ${body.headSha ?? "no-delivered-head"}`,
-              metadata: {
-                repairKind: failedRepair.repairKind,
-                repairTaskId: run.taskId,
-                startHeadSha: failedRepair.headSha,
-                targetHeadSha: failedRepair.baseHeadSha,
-                resolvedHeadSha: body.headSha ?? null,
-                state: "failed",
-              },
-            });
-            await openMergeTailStopNotice(tx, { taskId: failedRepair.regressionTaskId, agentId: run.agentId, sessionId: run.session.id, reason });
-          } else if (failedReview?.readinessTaskId && failedReview.regressionTaskId) {
-            const reason = `independent review ${run.taskId} failed without an exact-head decision for ${failedReview.headSha}`;
-            await tx.task.update({ where: { id: failedReview.readinessTaskId }, data: { status: TaskStatus.REVIEW, failureReason: reason } });
-            await tx.task.update({ where: { id: failedReview.regressionTaskId }, data: { status: TaskStatus.REVIEW, failureReason: reason } });
-            await openMergeTailStopNotice(tx, { taskId: failedReview.regressionTaskId, agentId: run.agentId, sessionId: run.session.id, reason });
-          }
-        }
-        // §4.0 outcome branching. The executor's own fenced write is the only
-        // writer of a step-12 output: neither synthesis nor the metadata update
-        // may touch it, because a synthesized body would read as a merge that
-        // never happened.
-        if (succeeded && mechanical && run.task) {
-          releaseMergeLeaseTask = tailLeaseChainId;
-          const persisted = await tx.taskStepOutput.findUnique({
-            where: { taskId: run.taskId }, select: { kind: true, body: true },
-          });
-          const outcome = parseMergeResult(persisted);
-          if (outcome.outcome === "merged") {
-            await advanceTemplateTask(tx, run.taskId, run.id, process.env.FEISHU_DEFAULT_CHAT_ID ?? null, now, completionTaskStatus);
-          } else {
-            await recordIntegratorStop(tx, {
-              integratorTaskId: run.taskId,
-              condition: outcome.outcome === "stopped" ? outcome.condition : "missing-or-malformed-result",
-              evidence: outcome.outcome === "stopped" ? outcome.evidence : outcome.reason,
-              agentId: run.agentId,
-              sessionId: run.session.id,
-              sourceRunId: run.id,
-            });
-          }
-          await tx.taskActivity.create({ data: {
-            taskId: run.taskId,
-            actorType: "runner",
-            actorId: body.runnerId,
-            body: outcome.outcome === "merged"
-              ? `Run ${run.runNumber} merged the chain's pull request`
-              : `Run ${run.runNumber} stopped before merging`,
-            metadata: jsonValue({ exitCode: body.exitCode, mergeOutcome: outcome.outcome }),
-          } });
-        } else if (succeeded && run.task && (run.task.templateId || run.task.chainId || mergeTailAuxiliary)) {
-          // Body, runId, metadata, and commit binding describe one act of
-          // authorship and only move together through task_output. Completion
-          // validates that immutable binding; it never restamps an authored
-          // body or synthesizes a canonical step's deliverable.
-          let existingOutput = await tx.taskStepOutput.findUnique({ where: { taskId: run.taskId } });
-          const canonicalAgentStep = isCanonicalAgentStep(run.task.templateStep);
-          const requiresExplicitOutput = canonicalAgentStep
-            || run.task.templateStep?.outputKind === "regression-verification";
-          if (!existingOutput && !requiresExplicitOutput) {
-            await tx.taskStepOutput.create({ data: {
-              taskId: run.taskId,
-              runId: run.id,
-              kind: run.task.templateStep?.outputKind ?? "result",
-              body: body.output?.trim() || `Run ${run.runNumber} completed successfully.`,
-              metadata: jsonValue({ branch: body.branch ?? run.branch, headSha: body.headSha }),
-              commitSha: body.headSha ?? null,
-            } });
-          } else if (!canonicalAgentStep && existingOutput?.runId === run.id && body.headSha) {
-            // Legacy and noncanonical steps retain their prose-compatible
-            // completion-time binding. Canonical artifacts are immutable and
-            // must already name the delivered head when authored.
-            existingOutput = await tx.taskStepOutput.update({
-              where: { id: existingOutput.id }, data: { commitSha: body.headSha },
-            });
-          }
-          const outputRefusal = canonicalOutputRefusal(
-            run.task.templateStep,
-            existingOutput,
-            run.id,
-            body.headSha ?? null,
-          );
-          if (outputRefusal) {
-            await tx.task.update({
-              where: { id: run.taskId },
-              data: { status: TaskStatus.REVIEW, failureReason: outputRefusal },
-            });
-            await tx.taskActivity.create({ data: {
-              taskId: run.taskId,
-              actorType: "control-plane",
-              body: `Canonical task output refused: ${outputRefusal}`,
-              metadata: {
-                kind: "canonicalTaskOutput.refusal",
-                schemaVersion: 1,
-                runId: run.id,
-                reason: outputRefusal,
-              },
-            } });
-            if (run.task.templateStep?.outputKind === "regression-verification") {
-              releaseMergeLeaseTask = tailLeaseChainId;
-            }
-          }
-          canonicalOutputFailure = outputRefusal;
-        }
-        if (succeeded && mechanical) {
-          // Already branched above; the mechanical path owns its own advance.
-        } else if (succeeded && run.task?.templateId) {
-          if (canonicalOutputFailure) {
-            // The current Run succeeded as a process, but it did not publish a
-            // canonical deliverable bound to that Run and head. The REVIEW
-            // state written above is the terminal control-plane outcome.
-          } else if (run.task.templateStep?.outputKind === "regression-verification") {
-            const result = await handleRegressionCompletion(tx, {
-              task: run.task,
-              run: {
-                id: run.id,
-                agentId: run.agentId,
-                branch: body.branch ?? run.branch,
-                headSha: body.headSha ?? null,
-                sessionId: run.session.id,
-              },
-              now,
-            });
-            if (result === "advance") {
-              await advanceTemplateTask(tx, run.taskId, run.id, process.env.FEISHU_DEFAULT_CHAT_ID ?? null, now, completionTaskStatus);
-            } else {
-              releaseMergeLeaseTask = tailLeaseChainId;
-            }
-          } else {
-            await advanceTemplateTask(tx, run.taskId, run.id, process.env.FEISHU_DEFAULT_CHAT_ID ?? null, now, completionTaskStatus);
-          }
-        } else if (succeeded && run.task && (run.task.chainId || mergeTailAuxiliary)) {
-          let repairUnable = false;
-          let reviewRejected = false;
-          if (reviewMarker?.readinessTaskId && reviewMarker.regressionTaskId && reviewMarker.headSha) {
-            const readinessTaskId = reviewMarker.readinessTaskId;
-            const regressionTaskId = reviewMarker.regressionTaskId;
-            const reviewHeadSha = reviewMarker.headSha;
-            const reviewBaseSha = reviewMarker.baseSha;
-            // Review evidence is run-scoped even though TaskStepOutput is not.
-            // An earlier attempt's decision is not this attempt's decision, and
-            // a decision not bound to the reviewed head is not evidence about
-            // it: either one would let an unreviewed head through.
-            const persistedReview = await tx.taskStepOutput.findUnique({
-              where: { taskId: run.taskId }, select: { body: true, runId: true, commitSha: true },
-            });
-            const reviewOutput = persistedReview?.runId === run.id && persistedReview.commitSha === reviewHeadSha
-              ? persistedReview
-              : null;
-            const parsedReview = reviewOutput
-              ? parseIndependentReviewDecision(reviewOutput.body, reviewHeadSha)
-              : {
-                status: "invalid" as const,
-                reason: persistedReview
-                  ? `decision is bound to run ${persistedReview.runId ?? "none"} at ${persistedReview.commitSha ?? "no head"}, not this run at ${reviewHeadSha}`
-                  : "missing independent review output",
-              };
-            const reviewSessionId = run.session.id;
-            // A finding's own text reaches these reasons, so every one of them
-            // is bounded exactly where a client-supplied reason would be.
-            const stopTail = async (unbounded: string): Promise<void> => {
-              const reason = truncateFailureReason(unbounded, FAILURE_REASON_LIMIT);
-              await tx.task.update({ where: { id: readinessTaskId }, data: { status: TaskStatus.REVIEW, failureReason: reason } });
-              await tx.task.update({ where: { id: regressionTaskId }, data: { status: TaskStatus.REVIEW, failureReason: reason } });
-              await openMergeTailStopNotice(tx, { taskId: regressionTaskId, agentId: run.agentId, sessionId: reviewSessionId, reason });
-            };
-            if (parsedReview.status === "invalid") {
-              reviewRejected = true;
-              const reason = `independent review returned an unusable decision for ${reviewHeadSha}: ${parsedReview.reason}`;
-              await tx.task.update({ where: { id: run.taskId }, data: { status: TaskStatus.DONE, failureReason: reason } });
-              await tx.task.update({ where: { id: readinessTaskId }, data: { status: TaskStatus.REVIEW, failureReason: reason } });
-              await openMergeTailStopNotice(tx, { taskId: regressionTaskId, agentId: run.agentId, sessionId: run.session.id, reason });
-            } else if (parsedReview.decision.outcome !== "rejected") {
-              // Follow-up findings never hold the merge. Each becomes a backlog
-              // card, so the merge proceeds with the work recorded instead of
-              // with the finding lost.
-              const followUpCardIds: string[] = [];
-              let followUpRefusal: string | null = null;
-              const followUpAgentName = await mergeTailFixAgentName(tx, await tx.task.findUnique({
-                where: { id: regressionTaskId },
-                select: { projectId: true, chainId: true, templateId: true },
-              }));
-              for (const finding of parsedReview.decision.findings) {
-                const card = await createReviewFollowUpCard(tx, {
-                  projectId: run.task.projectId,
-                  repoId: run.task.repoId,
-                  agentName: followUpAgentName,
-                  reviewTaskId: run.taskId,
-                  headSha: reviewHeadSha,
-                  finding,
-                });
-                if ("refusal" in card) { followUpRefusal = card.refusal; break; }
-                followUpCardIds.push(card.taskId);
-              }
-              if (followUpRefusal) {
-                reviewRejected = true;
-                const reason = `independent review follow-up card could not be created: ${followUpRefusal}`;
-                await tx.task.update({ where: { id: run.taskId }, data: { status: TaskStatus.DONE, failureReason: reason } });
-                await stopTail(reason);
-              } else {
-                await writeMarker(tx, readinessTaskId, "reviewObligation", {
-                  actorType: "control-plane",
-                  body: followUpCardIds.length === 0
-                    ? `Independent review approved exact head ${reviewHeadSha}`
-                    : `Independent review accepted exact head ${reviewHeadSha} with ${followUpCardIds.length} follow-up card(s)`,
-                  metadata: {
-                    state: parsedReview.decision.outcome,
-                    reviewTaskId: run.taskId,
-                    headSha: reviewHeadSha,
-                    baseSha: reviewBaseSha,
-                    followUpCardIds,
-                  },
-                });
-              }
-            } else {
-              reviewRejected = true;
-              const summary = parsedReview.decision.blockingSummary;
-              // The whole history, not the recent-state window: the blocking-round
-              // ceiling counts every rejection this readiness ever took, and one
-              // pushed past the window by later activity would reset the count.
-              const prior = await readMarkerHistory(tx, readinessTaskId);
-              const round = prior.filter((marker) => (
-                marker.kind === "reviewObligation" && marker.state === "rejected"
-              )).length + 1;
-              await writeMarker(tx, readinessTaskId, "reviewObligation", {
-                actorType: "control-plane",
-                body: `Independent review rejected exact head ${reviewHeadSha} on blocking round ${round}`,
-                metadata: {
-                  state: "rejected",
-                  reviewTaskId: run.taskId,
-                  headSha: reviewHeadSha,
-                  baseSha: reviewBaseSha,
-                  summary,
-                  blockingRound: round,
-                },
-              });
-              await tx.task.update({ where: { id: run.taskId }, data: {
-                status: TaskStatus.DONE,
-                failureReason: truncateFailureReason(`independent review rejected: ${summary}`, FAILURE_REASON_LIMIT),
-              } });
-              const driftRecovery = await baseDriftRecoveryContext(
-                tx,
-                regressionTaskId,
-                undefined,
-                reviewMarker.recoverySourceStopId ?? "no-recovery-context",
-              );
-              if (driftRecovery) {
-                await stopBaseDriftRecoveryTail(tx, driftRecovery, "independent-review", `rejected ${reviewHeadSha}: ${summary}`);
-              } else if (round >= MAX_BLOCKING_REVIEW_ROUNDS) {
-                // The ceiling is an exception path, not an approval gate: three
-                // blocking rounds mean the repair loop is not converging, so the
-                // tail stops and says so instead of spending a fourth round.
-                await stopTail(`independent review rejected ${reviewHeadSha} on blocking round ${round} of ${MAX_BLOCKING_REVIEW_ROUNDS}; automatic repair is exhausted: ${summary}`);
-              } else if (!reviewBaseSha) {
-                await stopTail(`independent review rejected ${reviewHeadSha} but its base head is unavailable for automatic repair`);
-              } else {
-                // The chain mutex is held now; the Regression row read before it
-                // may already be stale, and a repair bound to a stale repository
-                // or target branch cannot hand off to the fresh Regression run.
-                const lockedRegression = await tx.task.findUnique({
-                  where: { id: regressionTaskId },
-                  select: { projectId: true, repoId: true, templateId: true, chainId: true, targetBranch: true },
-                });
-                // Claiming the park this review owns is what makes the repair
-                // automatic without overruling anyone: an operator who moved
-                // readiness out of it while the review ran keeps that decision,
-                // and no repair run starts behind their back.
-                const claimedPark = await tx.task.updateMany({
-                  where: { id: readinessTaskId, status: TaskStatus.REVIEW, failureReason: { startsWith: INDEPENDENT_REVIEW_OPEN_PREFIX } },
-                  data: { failureReason: `review-fix: automatic repair queued at ${reviewHeadSha}` },
-                });
-                if (!lockedRegression) {
-                  await stopTail(`independent review rejected ${reviewHeadSha} but its Regression task is gone`);
-                } else if (claimedPark.count !== 1) {
-                  await writeMarker(tx, readinessTaskId, "reviewObligation", {
-                    actorType: "control-plane",
-                    body: `Independent review rejected ${reviewHeadSha}, but readiness is no longer parked on that review; automatic repair was not started`,
-                    metadata: {
-                      state: "repair-skipped",
-                      reviewTaskId: run.taskId,
-                      headSha: reviewHeadSha,
-                    },
-                  });
-                } else {
-                  const repair = await createMergeTailRepairTask(tx, {
-                    regressionTask: { id: regressionTaskId, ...lockedRegression },
-                    sourceRun: { id: run.id, branch: run.branch },
-                    agentName: await mergeTailFixAgentName(tx, lockedRegression),
-                    repairKind: "review-fix",
-                    headSha: reviewHeadSha,
-                    baseHeadSha: reviewBaseSha,
-                    summary,
-                    now,
-                  });
-                  if ("refusal" in repair) {
-                    await stopTail(`independent review rejected ${reviewHeadSha} and automatic repair was refused: ${repair.refusal}`);
-                  } else {
-                    await tx.task.update({
-                      where: { id: readinessTaskId },
-                      data: { status: TaskStatus.REVIEW, failureReason: `review-fix: automatic repair ${repair.taskId} queued at ${reviewHeadSha}` },
-                    });
-                  }
-                }
-              }
-            }
-          }
-          if (repairMarker?.regressionTaskId) {
-            const repairOutput = await tx.taskStepOutput.findUnique({ where: { taskId: run.taskId }, select: { body: true } });
-            let reportedUnable = false;
-            let resolvedHeadSha = body.headSha ?? null;
-            if (repairMarker.repairKind === "refresh-conflict") {
-              const parsedResolver = parseResolverResult(repairOutput?.body);
-              const expectedStart = repairMarker.headSha;
-              const expectedTarget = repairMarker.baseHeadSha;
-              const bindingError = parsedResolver.status === "invalid"
-                ? parsedResolver.reason
-                : parsedResolver.result.startHeadSha !== expectedStart || parsedResolver.result.targetHeadSha !== expectedTarget
-                  ? "merge-resolver output is bound to stale start or target heads"
-                  : parsedResolver.result.outcome === "resolved" && parsedResolver.result.resolvedHeadSha !== body.headSha
-                    ? "merge-resolver output resolved head does not match the delivered run head"
-                    : null;
-              if (bindingError) {
-                repairUnable = true;
-                const reason = `refresh-conflict repair ${run.taskId} returned invalid output: ${bindingError}`;
-                await tx.task.update({ where: { id: run.taskId }, data: { status: TaskStatus.DONE, failureReason: reason } });
-                await tx.task.update({ where: { id: repairMarker.regressionTaskId }, data: { status: TaskStatus.REVIEW, failureReason: reason } });
-                await writeMarker(tx, repairMarker.regressionTaskId, "repairResult", {
-                  actorType: "control-plane",
-                  body: `Automatic refresh-conflict attempt stopped: ${reason}`,
-                  metadata: {
-                    repairKind: "refresh-conflict",
-                    repairTaskId: run.taskId,
-                    startHeadSha: expectedStart,
-                    targetHeadSha: expectedTarget,
-                    resolvedHeadSha: body.headSha ?? null,
-                    state: "invalid-output",
-                    reason: bindingError,
-                  },
-                });
-                await openMergeTailStopNotice(tx, { taskId: repairMarker.regressionTaskId, agentId: run.agentId, sessionId: run.session.id, reason });
-              } else if (parsedResolver.status === "ok") {
-                reportedUnable = parsedResolver.result.outcome === "unable";
-                resolvedHeadSha = parsedResolver.result.outcome === "resolved" ? parsedResolver.result.resolvedHeadSha : null;
-              }
-            }
-            // gate-fix and review-fix agents have no JSON wire contract; their
-            // successful delivered head is the completion evidence.
-            if (reportedUnable) {
-              repairUnable = true;
-              const reason = `${String(repairMarker.repairKind)} repair ${run.taskId} reported unable at ${String(repairMarker.headSha)}`;
-              await tx.task.update({ where: { id: run.taskId }, data: { status: TaskStatus.DONE, failureReason: reason } });
-              await tx.task.update({ where: { id: repairMarker.regressionTaskId }, data: { status: TaskStatus.REVIEW, failureReason: reason } });
-              await openMergeTailStopNotice(tx, { taskId: repairMarker.regressionTaskId, agentId: run.agentId, sessionId: run.session.id, reason });
-            } else if (!repairUnable) {
-              await writeMarker(tx, repairMarker.regressionTaskId, "repairResult", {
-                actorType: "control-plane",
-                body: `Automatic ${String(repairMarker.repairKind)} attempt completed: ${String(repairMarker.headSha)} -> ${body.headSha ?? "missing-head"}`,
-                metadata: {
-                  repairKind: repairMarker.repairKind,
-                  repairTaskId: run.taskId,
-                  startHeadSha: repairMarker.headSha,
-                  targetHeadSha: repairMarker.baseHeadSha,
-                  resolvedHeadSha,
-                },
-              });
-              if (repairDocumentationTask) {
-                await tx.task.update({
-                  where: { id: repairDocumentationTask.id },
-                  data: {
-                    status: TaskStatus.TODO,
-                    failureReason: `documentation invalidated by ${String(repairMarker.repairKind)} repair ${run.taskId}`,
-                  },
-                });
-              }
-            }
-          }
-          if (repairUnable || reviewRejected) {
-            // The failed repair owns the stop; never activate its follow-up.
-            releaseMergeLeaseTask = tailLeaseChainId;
-          } else if (run.task.approvalGate) {
-            const claimed = await tx.task.updateMany({
-              where: { id: run.taskId, status: completionTaskStatus! },
-              data: { status: TaskStatus.REVIEW, failureReason: null },
-            });
-            if (claimed.count === 1) await gateQuestion(tx, run.taskId, run.id, process.env.FEISHU_DEFAULT_CHAT_ID ?? null);
-          } else {
-            const completed = await tx.task.updateMany({
-              where: { id: run.taskId, status: completionTaskStatus! }, data: { status: TaskStatus.DONE, failureReason: null },
-            });
-            if (completed.count === 1) {
-              if (run.task.chainId) {
-                await activateChainSuccessor(tx, run.task, {
-                  sourceRunId: run.id,
-                  chatId: process.env.FEISHU_DEFAULT_CHAT_ID ?? null,
-                }, now);
-              }
-              if (auxiliaryTargetTaskId) await activateMergeTailTarget(tx, auxiliaryTargetTaskId, now);
-            }
-          }
-        } else {
-          await tx.task.updateMany({
-            where: { id: run.taskId, ...(completionTaskStatus ? { status: completionTaskStatus } : {}) },
-            data: {
-              status: retryCreated ? TaskStatus.DOING : TaskStatus.REVIEW,
-              failureReason: succeeded ? null : budgetExhausted
-                ? `Maximum ${budgetCeiling} runs reached`
-                : body.failureReason ?? "Execution failed",
-            },
-          });
-        }
-        if (!(succeeded && mechanical)) await tx.taskActivity.create({
-          data: {
-            taskId: run.taskId,
-            actorType: "runner",
-            actorId: body.runnerId,
-            body: canonicalOutputFailure ? `Run ${run.runNumber} succeeded but canonical task output was refused`
-              : succeeded && (run.task?.templateId || run.task?.chainId || mergeTailAuxiliary) ? `Run ${run.runNumber} succeeded; chain advanced or awaiting approval`
-              : succeeded ? `Run ${run.runNumber} succeeded; task moved to review`
-              : retryCreated ? `Run ${run.runNumber} failed; retry queued`
-                : `Run ${run.runNumber} failed; task moved to review`,
-            metadata: jsonValue({ exitCode: body.exitCode, terminalEventSeen: body.terminalEventSeen, failureClass, pushStatus: body.pushStatus, pullRequestUrl: body.pullRequestUrl }),
-          },
-        });
-        if (budgetExhausted) {
-          await tx.inboxMessage.create({
-            data: {
-              from: "AGENT",
-              sessionId: run.session.id,
-              taskId: run.taskId,
-              kind: "TEXT",
-              body: `Run budget exhausted after ${budgetCeiling} attempts; operator action required.`,
-            },
-          });
-        }
-      }
-      if (failureClass === FailureClass.AUTH_REQUIRED) {
-        const state = await tx.runnerBackendState.upsert({
-          where: { runner: run.runner },
-          create: { runner: run.runner, consecutiveAuthFailures: 1, lastPreflightOk: false },
-          update: { consecutiveAuthFailures: { increment: 1 }, lastPreflightOk: false },
-        });
-        if (state.consecutiveAuthFailures >= 2) {
-          await tx.runnerBackendState.update({
-            where: { runner: run.runner },
-            data: { circuitOpen: true, circuitReason: "Repeated authentication failures", circuitOpenedAt: now },
-          });
-          await tx.inboxMessage.create({
-            data: {
-              from: "AGENT",
-              sessionId: run.session.id,
-              taskId: run.taskId,
-              goalId: run.goalId,
-              kind: "TEXT",
-              body: `${run.runner.toLowerCase()} runner circuit opened after repeated authentication failures; login is required.`,
-            },
-          });
-        }
-      } else if (succeeded) {
-        await tx.runnerBackendState.upsert({
-          where: { runner: run.runner },
-          create: { runner: run.runner, lastPreflightOk: true },
-          update: { consecutiveAuthFailures: 0 },
-        });
-      }
-      return { taskId: run.taskId, succeeded, retryCreated, failureClass, releaseMergeLeaseTask };
-    // ReadCommitted lets successor CAS losers observe count=0 instead of
-    // surfacing a serialization failure to runners. Every task status write
-    // above has its own status CAS so concurrent operator decisions win.
-    }, { isolationLevel: Prisma.TransactionIsolationLevel.ReadCommitted });
-    if (!result) {
-      const waiting = await db.run.findFirst({ where: { id: runId, status: RunStatus.WAITING_INBOX }, select: { id: true } });
-      return waiting
+    if ("kind" in result) {
+      if (result.kind === "principal") return context.json({ error: result.error }, 403);
+      return result.kind === "waiting-inbox"
         ? context.json({ error: "Run suspended for Inbox", code: "WAITING_INBOX" }, 409)
-        : context.json({ error: "Stale fencing token", reason: await explainFenceRefusal(db, fence) }, 409);
+        : context.json({ error: "Stale fencing token", reason: result.reason }, 409);
     }
     await releaseMergeLeaseSafely(releaseChainLease, result.releaseMergeLeaseTask);
     await options.ownership.assertHeld();
